requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
flask-cors==4.0.1
flask-talisman==1.1.0
prometheus-flask-exporter==0.23.0
gunicorn==22.0.0
